kiosk list 출력 시 category 값 기반 출력 -> 키오스크 창으로 이동하는 기능 수정 필요 (각 사용자에 맞는 단계 호출되도록)
