+++ src/main/java/com/operation/dao/KioskDAO.java
@@ -6,6 +6,7 @@ import org.apache.ibatis.session.SqlSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
+import com.operation.dto.KioskCategoryDTO;
 import com.operation.dto.KioskDTO;
 import com.operation.dto.KioskRecordDTO;
 
@@ -15,9 +16,14 @@ public class KioskDAO {
 	@Autowired
 	private SqlSession db;
 	
-	// 키오스크 목록 불러오기
-	public List<KioskDTO> selectAll(int is_game){
-		return db.selectList("Kiosk.selectAll", is_game);
+//	// 키오스크 목록 불러오기
+//	public List<KioskDTO> selectAll(int is_game){
+//		return db.selectList("Kiosk.selectAll", is_game);
+//	}
+	
+	// 키오스크 카테고리 목록 불러오기
+	public List<KioskCategoryDTO> selectAll(){
+		return db.selectList("Kiosk.selectAll");
 	}
 	
 	// 키오스크 상세 정보 불러오기
