+++ src/main/java/com/operation/dto/KioskDTO.java
@@ -3,7 +3,6 @@ package com.operation.dto;
 public class KioskDTO {
 	private int id;
 	private String kiosk_category_id;
-	private String name;
 	private String url;
 	private int play_stage;
 	private int width;
@@ -14,11 +13,10 @@ public class KioskDTO {
 		super();
 	}	
 
-	public KioskDTO(int id, String kiosk_category_id, String name, String url, int play_stage, int width, int height,
+	public KioskDTO(int id, String kiosk_category_id, String url, int play_stage, int width, int height,
 			boolean is_game) {
 		this.id = id;
 		this.kiosk_category_id = kiosk_category_id;
-		this.name = name;
 		this.url = url;
 		this.play_stage = play_stage;
 		this.width = width;
@@ -40,14 +38,6 @@ public class KioskDTO {
 
 	public void setKiosk_category_id(String kiosk_category_id) {
 		this.kiosk_category_id = kiosk_category_id;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
 	}
 
 	public String getUrl() {
